[PureGo] Follow-up review cleanup: dedup requeue, test-only helper, tidy layout
- Remove the redundant buf.requeue() calls in supervisor (retry-backoff and
  pause paths). runOnce already requeues on every successful Open, so the
  supervisor's calls were no-ops in flight and misleading in intent.
- Move the test-only buffer.len helper out of production code into an
  export_test.go file.
- Return NewEphemeralOpener as opener[encodedMsg, ephemeralResp] (interface)
  rather than the unexported concrete *ephemeralOpener, matching Go idiom for
  a factory constructor.
- Move ack-model unit tests from encoder_test.go to their own
  ackmodel_test.go so tests line up with the file they cover.
- Refresh architecture.md status: transport + auth (incl. UC OAuth) + generic
  proto/JSON core are built; only the public zerobus package and the Arrow
  wire path remain. Replace the "edge #1/#2/#3" internal-numbering references
  with self-contained descriptions.
